Fix Color Variations, Create New Screenshot
(1) Some of the colors between screenshots that were supposed to be the
same colors were close matches, but not exact. This appears to be due to
using the color picker tool in Boxy SVG to select colors from the
rendered websites. Although the tool selected a very similar color to
the color rendered on the websites, it did not match the actual HSL
colors that were used in the CSS. Therefore, all colors in every
screenshot were inspected and changed if necessary to match the CSS
colors.

(2) Screenshot-3 was created to diagram a new function that allows users
to remove hidden jobs from display.