Always-Enabled Action/Popup
(1) Action/Popup is now always enabled. Previously, it was disabled on
tabs that did not match a supported job board's URL pattern. Now, if a
tab's URL does not match a job board's URL pattern, clicking the action
button will open a popup that allows users to search for jobs on
supported job boards, such as LinkedIn and Indeed.

(2) Content-scripts are now injected as long as there is a domain name
match, without consideration of the path. This change was made for the
following reasons: (a) allow Hide n' Seek to work on job feeds (Indeed)
and job collections (LinkedIn), and (b) to fix content-script injection
issues related to back/forward navigation on LinkedIn, which seems to
work like a single page application (SPA) website.

(3) Installation of and updates to Hide n' Seek will now cause relevant
tabs to refresh in order to flush out disconnected content-scripts, as
well as to inject new content-scripts.