Update JobRegistrar
(1) The JobRegistrar's mutation observer's root is now the html element.
This change was made due to issues with back/forward navigation on
LinkedIn, which works like a single page application (SPA), causing
removal of the mutation observer's root element, "jobElementsContainer".

(2) The JobRegistrar's mutation observer callback now processes
addedNodes instead of just targets. This allows it to catch
dynamically-loaded jobs on Indeed's job feed.

(3) A new method, startJobRegisteredBeacon, was added. It accomplishes
two things: (a) it tells the background script when a content script has
been loaded, in case the popup needs to change from inapplicable-tab
status to applicable-tab status, and (b) it responds to the popup script
when it requests the jobBoardId from the content-script in the active
tab within the current window.

(4) An additional "jobElement" selector was added to JobRegistrar's
selectors constructor object for LinkedIn. This addition allows
JobRegistrar's mutation observer to catch job elements on LinkedIn's
"collections" pages. Additionally, a new selector for "companyName" was
added to the selectors object in the JobElementDataGetter class for
LinkedIn.

(5) ondragstart event handlers were added to dynamically-loaded img
elements with preventDefault called to prevent dragging.

(6) jobBoardId is now set as a data-attribute on jobBlockElements.
Currently, this allows the stylesheet to make adjustments when necessary
based on the job board.