Add Rules for Inapplicable Tab Popup
(1) A new set of rules for the inapplicable tab popup have been added.

(2) popup html height is now determined by the value of a data-attribute
called data-applicable-tab.

(3) Some div tags were converted to button tags, and rules were adjusted
accordingly to account for the user-agent's default values for buttons.

(4) data-list-is-scrollable is no longer necessary, so rules for it were
removed

(5) Keyboard accessibility is improved. However, the use of the :has()
pseudo-class may be reconsidered due to Firefox not currently supporting
its use.

(6) The "toggle-container" class's opacity was increased.