Add InapplicableTabPopup Class
(1) The InapplicableTabPopup class was added to handle popup behaviors
when the action button is clicked on tabs that do not have a jobBoardId.

(2) ondragstart event handlers were added to dynamically-created img
elements along with a call to preventDefault in order to block dragging.

(3) A keydown event listener was added to the "Do not display hidden
jobs" checkbox input element to catch "Enter" keys. By default, it only
catches "Space". The keydown handler simply dispatches an input event
to the checkbox's input handler.

(4) jobBoardIds are now retrieved by sending a message as a request to
the active tab in the current window, rather than relying on URL match
patterns. This lets the popup know whether to show the "applicable tab"
variation, or the "inapplicable tab" variation of the popup. An
applicable tab is considered to be a tab that has a content script which
has successfully registered at least one job. A tab with a
content-script in this state will reply with the jobBoardId, whereas
other tabs will not be able to.