Revise Action/Badge Update Logic
(1) Action and badge updates are now primarily driven by content script
status changes and storage changes. The previous implementation had
issues with single page application (SPA) websites. For example, on
LinkedIn, the popup should show the "applicable tab" version of the
popup. However, if a user were to navigate to a non-job-search section
of LinkedIn, the popup should still show the "applicable tab" version.
This is because the content script does not reload on SPA websites.
This is no longer an issue because now the content script monitors the
status of the Hide n' Seek UI and updates the background and popup
scripts when there are changes so that the action button and popup can
show the correct information.