Update Logic Order of registerJobs Method
(1) The private property, #hasHideNSeekUI, of the JobRegistrar class is
now updated before sending a message to the background script. This is
because when the background script receives the message, it will get
the hasHideNSeekUI statuses of all tabs, including the tab that sent
the initial message, so the hasHideNSeekUI property should be updated
before the message is sent to the background script.