Update JobBoardSelectors Class
(1) The jobElement selector is now created from selector components
instead of being in the selectors literal object.

(2) The card selector type has been added. It used to be part of the
jobElement type.

(3) The card selector for the signed-out version of LinkedIn has been
updated from ".base-card" to ".job-search-card" for improved
specificity. The use of the ".base-card" selector was resulting in
false positive job listings on the signed-out version of LInkedin.